# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Tracker API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/task-flow/stages-stats": {
            "get": {
                "tags": [
                    "Task Flow"
                ],
                "summary": "Stage statistics (counts per pipeline stage)",
                "operationId": "f606d06902ed6868f25a47f07505cbe0",
                "parameters": [
                    {
                        "name": "mine",
                        "in": "query",
                        "description": "Only tasks where current user is responsible",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task-flow/sub-stats": {
            "get": {
                "tags": [
                    "Task Flow"
                ],
                "summary": "Checkpoint status counts for a stage",
                "operationId": "532993f18c2598e39ca081128889b21e",
                "parameters": [
                    {
                        "name": "stage_id",
                        "in": "query",
                        "description": "Stage identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mine",
                        "in": "query",
                        "description": "Only tasks where current user is responsible",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Stage not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task-flow/collection": {
            "get": {
                "tags": [
                    "Task Flow"
                ],
                "summary": "Filtered task list (paginated)",
                "operationId": "1471b84533c37bb083176acd0ed31b71",
                "parameters": [
                    {
                        "name": "stage_id",
                        "in": "query",
                        "description": "Filter by stage",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "checkpoint_status",
                        "in": "query",
                        "description": "Filter by checkpoint status: pending, claimed, in_progress, done",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mine",
                        "in": "query",
                        "description": "Only user responsible tasks",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}